fix: resolved compose conflict

+++ app/src/main/java/com/denariidolor/presentation/ui/settings/SettingsScreenState.kt
@@ -1,6 +1,6 @@
 package com.denariidolor.presentation.ui.settings
 
 data class SettingsScreenState(
-    val sessionTimeoutMinutes: Int,
+    val sessionTimeoutMinutes: Long,
     val pinConfigured: Boolean
 )
